Refactor: Split thread color management into AddEmbroideryThreadColorForm and AddStitchingThreadColorForm components, move threadColorUtils to app/lib/utils, clean up app.addThreadColors.jsx, and update .gitignore for Cursor/specstory files
