+++ app/routes/app.jsx
@@ -38,8 +38,8 @@ export default function App() {
         </Link>
         <Link to="/app/createProducts">Create a new product</Link>
         <Link to="/app/addLeatherColor">Add Leather Color</Link>
-        <Link to="/app/updatePricing">Update Pricing</Link>
         <Link to="/app/addThreadColors">Add Thread Colors</Link>
+        <Link to="/app/updatePricing">Update Pricing</Link>
       </NavMenu>
       <Outlet />
     </PolarisProvider>
