+++ shopify.app.dev.toml
@@ -3,7 +3,7 @@
 client_id = "df58a5b13484ffc9b63eb3e420dca6f9"
 name = "LPC Product Management - Dev"
 handle = "lpc-product-management-dev-1"
-application_url = "https://sense-remedies-twisted-asbestos.trycloudflare.com"
+application_url = "https://drink-taxi-simplified-screensaver.trycloudflare.com"
 embedded = true
 
 [build]
@@ -17,9 +17,9 @@ scopes = "read_fulfillments,read_inventory,read_locations,read_orders,read_produ
 
 [auth]
 redirect_urls = [
-  "https://sense-remedies-twisted-asbestos.trycloudflare.com/auth/callback",
-  "https://sense-remedies-twisted-asbestos.trycloudflare.com/auth/shopify/callback",
-  "https://sense-remedies-twisted-asbestos.trycloudflare.com/api/auth/callback"
+  "https://drink-taxi-simplified-screensaver.trycloudflare.com/auth/callback",
+  "https://drink-taxi-simplified-screensaver.trycloudflare.com/auth/shopify/callback",
+  "https://drink-taxi-simplified-screensaver.trycloudflare.com/api/auth/callback"
 ]
 
 [webhooks]
